refactor(api): split brain.ts into per-resource sub-routers (A17.3)
The 1734-line api/brain.ts with 39 routes is now api/brain/:
  memories.ts (6), entities.ts (9), edges.ts (6), chrono.ts (7),
  file-meta.ts (3), search.ts (7: stats/traverse/query/recall/find-similar/
  reindex), bulk.ts (1), over a _shared.ts with the pieces every sub-router
  needs (webhookToken, getSpaceMeta, applyValidation, buildMemoryFilter,
  UUID_V4_RE). index.ts mounts them on the same brainRouter.

Each sub-router declares full paths, so every URL is unchanged, and original
route order is preserved within each file (entities /by-name and /by-ids must
stay ahead of /:id). Handlers moved verbatim.

Two notes:
- The brain file-metadata router is fileMetaRouter, not filesRouter: api/files.ts
  already exports a filesRouter, and the duplicate name made the audit-coverage
  guard resolve those routes to the wrong (/api/files) prefix.
- The two route-analysis guards (audit-route-coverage, route-guard-coverage) now
  walk the api dir recursively and resolve sub-routers mounted via
  parentRouter.use(child) to the parent's prefix — without that the whole brain
  surface silently drops out of both checks. They caught this refactor; keeping
  them honest is the point.

Verified: build clean, no dead imports (tsc --noUnusedLocals); brain 183/183,
mcp-tools 94/94, proxy-spaces 44/44, standalone 800/0, sync 191/0,
red-team 272/0.
