--- server/src/api/brain.ts
@@ -1,1734 +1,0 @@
-import { Router } from 'express';
-import { toDocId } from '../util/paths.js';
-import { escapeRegex } from '../util/redos.js';
-import type express from 'express';
-import { v4 as uuidv4 } from 'uuid';
-import { requireSpaceAuth, denyReadOnly } from '../auth/middleware.js';
-import { globalRateLimit, bulkWipeRateLimit } from '../rate-limit/middleware.js';
-import { NotFoundError } from '../util/errors.js';
-import { listMemories, deleteMemory, countMemories, bulkDeleteMemories, remember, updateMemory, queryBrain, findSimilar, recall, validateFilterExpression, type RecallKnowledgeType, type FilterExpression } from '../brain/memory.js';
-import { listEntities, deleteEntity, upsertEntity, getEntityById, updateEntityById, bulkDeleteEntities, findEntitiesByName, findEntityBacklinks } from '../brain/entities.js';
-import { listEdges, deleteEdge, upsertEdge, getEdgeById, updateEdgeById, bulkDeleteEdges, traverseGraph, traverseRecallSeeds, MAX_RECALL_TRAVERSE, resolveEdgeEntityNames } from '../brain/edges.js';
-import { memoryEmbedText, entityEmbedText, edgeEmbedText, chronoEmbedText, fileEmbedText } from '../brain/embed-text.js';
-import { bulkWrite, bulkWriteTotal, type BulkInput } from '../brain/bulk.js';
-import { computeMergePlan, applyResolutions, executeMerge, validateResolution, type PropertyResolution } from '../brain/merge.js';
-import { validateDeleteFields, applyDeleteFields as applyDeleteFieldsPaths } from '../brain/delete-fields.js';
-/** Regex that matches a UUID v4 (case-insensitive). */
-const UUID_V4_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
-import { createChrono, updateChrono, getChronoById, listChrono, deleteChrono, bulkDeleteChrono, parseRecurrence, ChronoFilter } from '../brain/chrono.js';
-import { embed } from '../brain/embedding.js';
-import { updateFileMeta, deleteFileMeta } from '../files/file-meta.js';
-import { fileExists } from '../files/files.js';
-import { getConfig } from '../config/loader.js';
-import { col, asFilter, asDoc } from '../db/mongo.js';
-import { nextSeq } from '../util/seq.js';
-import { needsReindex, clearReindexFlag } from '../spaces/spaces.js';
-import { log } from '../util/log.js';
-import { parseLimit, parseSkip, capPage } from '../util/pagination.js';
-import { checkQuota, QuotaError } from '../quota/quota.js';
-import { resolveMemberSpaces, resolveWriteTarget, findSpace, isProxySpace, isStrictLinkage, findFirstAcrossMembers, collectAcrossMembers } from '../spaces/proxy.js';
-import { validateEntity, validateEdge, validateMemory, validateChrono, resolveMetaRefs, getAllowedChronoTypes, type SchemaViolation } from '../spaces/schema-validation.js';
-import type { MemoryDoc, EntityDoc, EdgeDoc, ChronoEntry, FileMetaDoc, ChronoType, ChronoStatus, SpaceMeta } from '../config/types.js';
-import { reindexInProgress } from '../metrics/registry.js';
-import { emitWebhookEvent } from '../webhooks/dispatcher.js';
-
-export const brainRouter = Router();
-let reindexJobRunning = false;
-
-// ── Webhook helper ────────────────────────────────────────────────────────
-
-/** Extract token identification from the request for webhook payloads. */
-function webhookToken(req: express.Request): { tokenId?: string; tokenLabel?: string } {
-  const t = req.authToken;
-  if (!t) return {};
-  return {
-    tokenId: 'id' in t ? (t as { id: string }).id : undefined,
-    tokenLabel: t.name,
-  };
-}
-
-// ── Schema validation helpers ─────────────────────────────────────────────
-
-/** Look up the meta block for a space from config, with library refs resolved. Returns undefined if none. */
-function getSpaceMeta(spaceId: string): SpaceMeta | undefined {
-  const cfg = getConfig();
-  const meta = cfg.spaces.find(s => s.id === spaceId)?.meta;
-  if (!meta) return undefined;
-  return resolveMetaRefs(meta);
-}
-
-/**
- * Apply schema validation to a write operation.
- * Returns { blocked: true, violations } when strict mode rejects the write.
- * Returns { blocked: false, warnings } when warn mode lets the write through.
- * Returns { blocked: false, warnings: [] } when validation is off or no meta.
- */
-function applyValidation(
-  meta: SpaceMeta | undefined,
-  violations: SchemaViolation[],
-): { blocked: boolean; warnings: SchemaViolation[] } {
-  if (!meta || !meta.validationMode || meta.validationMode === 'off' || violations.length === 0) {
-    return { blocked: false, warnings: [] };
-  }
-  if (meta.validationMode === 'strict') {
-    return { blocked: true, warnings: violations };
-  }
-  // warn mode
-  return { blocked: false, warnings: violations };
-}
-
-// ── Short-form memory CRUD  (/:spaceId/memories) ──────────────────────────
-// These are the primary REST endpoints used by API clients and integration tests.
-
-// POST /api/brain/spaces/:spaceId/memories — create a memory
-brainRouter.post('/spaces/:spaceId/memories', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  // Proxy space: resolve target space for write
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const targetSpace = wt.target;
-  const { fact, tags = [], entityIds = [], description, properties, type: memoryType } = req.body ?? {};
-  if (!fact || typeof fact !== 'string') {
-    res.status(400).json({ error: '`fact` string required' });
-    return;
-  }
-  if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) {
-    res.status(400).json({ error: '`tags` must be an array of strings' });
-    return;
-  }
-  if ((fact as string).length > 50_000) {
-    res.status(400).json({ error: '`fact` must not exceed 50 000 characters' });
-    return;
-  }
-  // Quota check — reject with 507 if brain hard limit exceeded
-  let quotaResult;
-  try {
-    quotaResult = await checkQuota('brain');
-  } catch (err) {
-    if (err instanceof QuotaError) {
-      res.status(507).json({ error: err.message, storageExceeded: true });
-      return;
-    }
-    throw err;
-  }
-  const safeDesc: string | undefined = typeof description === 'string' ? description : undefined;
-  const safeProps: Record<string, string | number | boolean> | undefined =
-    properties != null && typeof properties === 'object' && !Array.isArray(properties)
-      ? (properties as Record<string, string | number | boolean>)
-      : undefined;
-  const safeEntityIds: string[] = Array.isArray(entityIds) ? entityIds : [];
-  // Validate that all entityIds are valid UUID v4 (not names) — only when strictLinkage is on
-  if (isStrictLinkage(wt.target)) {
-    const invalidEntityIds = safeEntityIds.filter((id: string) => !UUID_V4_RE.test(id));
-    if (invalidEntityIds.length > 0) {
-      res.status(400).json({ error: '`entityIds` must contain valid UUID v4 values (entity IDs), not names', invalid: invalidEntityIds });
-      return;
-    }
-  }
-  const safeTags: string[] = Array.isArray(tags) ? tags : [];
-
-  // Schema validation
-  const safeMemoryType: string | undefined = typeof memoryType === 'string' ? memoryType : undefined;
-  const meta = getSpaceMeta(wt.target);
-  const violations = validateMemory(meta ?? {}, { type: safeMemoryType, properties: safeProps });
-  const validation = applyValidation(meta, violations);
-  if (validation.blocked) {
-    res.status(400).json({ error: 'schema_violation', violations: validation.warnings });
-    return;
-  }
-
-  // Persist through the shared remember() so REST and MCP produce identical records: the same
-  // embed-text derivation (properties folded as `key value` via propsEmbedText, entity names
-  // resolved consistently), the `matchedText` source string, the author, insert-time duplicate-
-  // rule evaluation, and the memory.created webhook. Previously inlined here, which had drifted
-  // into three bugs: values-only property embedding, no `matchedText`, and no dupe-rule firing.
-  const doc = await remember(
-    targetSpace, fact, safeEntityIds, safeTags, safeDesc, safeProps,
-    undefined, safeMemoryType, undefined, webhookToken(req),
-  );
-  const body: Record<string, unknown> = { ...doc };
-  if (quotaResult?.softBreached) body['storageWarning'] = true;
-  if (validation.warnings.length > 0) body['warnings'] = validation.warnings;
-  res.status(201).json(body);
-});
-
-/** Build a MongoDB filter from `tag` and `entity` query params */
-function buildMemoryFilter(query: Record<string, unknown>): Record<string, unknown> {
-  const filter: Record<string, unknown> = {};
-  const tag = typeof query['tag'] === 'string' ? query['tag'] : undefined;
-  const entity = typeof query['entity'] === 'string' ? query['entity'] : undefined;
-  const type = typeof query['type'] === 'string' ? query['type'] : undefined;
-  if (tag) filter['tags'] = { $regex: `^${escapeRegex(tag)}$`, $options: 'i' };
-  if (entity) filter['entityIds'] = entity;
-  if (type) filter['type'] = type;
-  return filter;
-}
-
-// GET /api/brain/spaces/:spaceId/memories/:id — get single memory
-brainRouter.get('/spaces/:spaceId/memories/:id', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const doc = await findFirstAcrossMembers(spaceId,
-    mid => col<MemoryDoc>(`${mid}_memories`).findOne(asFilter<MemoryDoc>({ _id: id })));
-  if (doc) { res.json(doc); return; }
-  res.status(404).json({ error: 'Memory not found' });
-});
-
-// GET /api/brain/spaces/:spaceId/stats
-brainRouter.get('/spaces/:spaceId/stats', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const memberIds = resolveMemberSpaces(spaceId);
-  const counts = await Promise.all(memberIds.map(async mid => ({
-    memories: await countMemories(mid),
-    entities: await col(`${mid}_entities`).countDocuments(),
-    edges: await col(`${mid}_edges`).countDocuments(),
-    chrono: await col(`${mid}_chrono`).countDocuments(),
-    // Exclude chunk records (parentFileId set) — count only top-level file records
-    files: await col(`${mid}_files`).countDocuments({ parentFileId: { $exists: false } }),
-  })));
-  const memories = counts.reduce((s, c) => s + c.memories, 0);
-  const entities = counts.reduce((s, c) => s + c.entities, 0);
-  const edges = counts.reduce((s, c) => s + c.edges, 0);
-  const chrono = counts.reduce((s, c) => s + c.chrono, 0);
-  const files = counts.reduce((s, c) => s + c.files, 0);
-  res.json({ spaceId, memories, entities, edges, chrono, files });
-});
-
-// GET /api/brain/spaces/:spaceId/memories
-brainRouter.get('/spaces/:spaceId/memories', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const limit = parseLimit(req.query['limit'], 100, 500);
-  const skip = parseSkip(req.query['skip']);
-  const filter = buildMemoryFilter(req.query as Record<string, unknown>);
-  const all = await collectAcrossMembers(spaceId, mid => listMemories(mid, filter, limit, skip));
-  res.json({ memories: capPage(all, limit), limit, skip });
-});
-
-// DELETE /api/brain/spaces/:spaceId/memories/:id
-brainRouter.delete('/spaces/:spaceId/memories/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const deleted = await findFirstAcrossMembers(spaceId, mid => deleteMemory(mid, id, webhookToken(req)));
-  if (deleted) { res.status(204).end(); return; }
-  res.status(404).json({ error: 'Memory not found' });
-});
-
-// PATCH /api/brain/spaces/:spaceId/memories/:id — partial update a memory (long-form)
-brainRouter.patch('/spaces/:spaceId/memories/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const { fact, tags, entityIds, description, properties, deleteFields } = req.body ?? {};
-  // Validate deleteFields
-  const dfResult = validateDeleteFields(deleteFields);
-  if (!dfResult.ok) { res.status(400).json({ error: dfResult.error }); return; }
-  const dfPaths: string[] | undefined = Array.isArray(deleteFields) && deleteFields.length > 0 ? deleteFields : undefined;
-  const updates: { fact?: string; tags?: string[]; entityIds?: string[]; description?: string; properties?: Record<string, string | number | boolean> } = {};
-  if (fact !== undefined) {
-    if (typeof fact !== 'string' || !fact.trim()) { res.status(400).json({ error: '`fact` must be a non-empty string' }); return; }
-    updates.fact = fact;
-  }
-  if (tags !== undefined) {
-    if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) { res.status(400).json({ error: '`tags` must be an array of strings' }); return; }
-    updates.tags = tags;
-  }
-  if (entityIds !== undefined) {
-    if (!Array.isArray(entityIds) || entityIds.some((t: unknown) => typeof t !== 'string')) { res.status(400).json({ error: '`entityIds` must be an array of strings' }); return; }
-    if (isStrictLinkage(wt.target)) {
-      const invalidIds = entityIds.filter((id: string) => !UUID_V4_RE.test(id));
-      if (invalidIds.length > 0) { res.status(400).json({ error: '`entityIds` must contain valid UUID v4 values (entity IDs), not names', invalid: invalidIds }); return; }
-    }
-    updates.entityIds = entityIds;
-  }
-  if (description !== undefined) {
-    if (typeof description !== 'string') { res.status(400).json({ error: '`description` must be a string' }); return; }
-    updates.description = description;
-  }
-  if (properties !== undefined) {
-    if (typeof properties !== 'object' || properties === null || Array.isArray(properties)) { res.status(400).json({ error: '`properties` must be a plain object' }); return; }
-    updates.properties = properties as Record<string, string | number | boolean>;
-  }
-  if (Object.keys(updates).length === 0 && !dfPaths) { res.status(400).json({ error: 'At least one field must be provided' }); return; }
-  const memberIds = resolveMemberSpaces(wt.target);
-  for (const mid of memberIds) {
-    // Schema validation after deleteFields + merge for memories
-    if (dfPaths) {
-      const existing = await listMemories(mid, { _id: id }, 1, 0);
-      if (existing.length === 0) continue;
-      const mem = existing[0]!;
-      const resultProps = updates.properties ?? (mem.properties != null ? { ...mem.properties } : {});
-      const sim: Record<string, unknown> = { properties: resultProps };
-      applyDeleteFieldsPaths(sim, dfPaths);
-      const simProps = (sim['properties'] ?? {}) as Record<string, unknown>;
-      const meta = getSpaceMeta(mid);
-      const violations = validateMemory(meta ?? {}, { properties: simProps });
-      const validation = applyValidation(meta, violations);
-      if (validation.blocked) {
-        res.status(422).json({ error: 'schema_violation', message: 'deleteFields + merge result violates required properties', violations: validation.warnings });
-        return;
-      }
-    }
-    const updated = await updateMemory(mid, id, updates, dfPaths, webhookToken(req));
-    if (updated) {
-      res.json(updated);
-      return;
-    }
-  }
-  res.status(404).json({ error: 'Memory not found' });
-});
-
-// DELETE /api/brain/spaces/:spaceId/memories — bulk wipe (long-form)
-brainRouter.delete('/spaces/:spaceId/memories', bulkWipeRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  if (isProxySpace(spaceId)) {
-    res.status(400).json({ error: 'Bulk wipe not supported on proxy spaces — target member spaces individually' });
-    return;
-  }
-  if (req.body?.confirm !== true) {
-    res.status(400).json({ error: '`confirm: true` required in request body' });
-    return;
-  }
-  const deleted = await bulkDeleteMemories(spaceId);
-  res.json({ deleted });
-});
-
-// POST /api/brain/spaces/:spaceId/entities — create/upsert an entity
-brainRouter.post('/spaces/:spaceId/entities', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const { id, name, type = '', tags = [], properties = {}, description } = req.body ?? {};
-  if (id !== undefined) {
-    if (typeof id !== 'string' || !UUID_V4_RE.test(id)) {
-      res.status(400).json({ error: '`id` must be a valid UUID v4' });
-      return;
-    }
-  }
-  if (!name || typeof name !== 'string') {
-    res.status(400).json({ error: '`name` string required' });
-    return;
-  }
-  if (typeof type !== 'string') {
-    res.status(400).json({ error: '`type` must be a string' });
-    return;
-  }
-  if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) {
-    res.status(400).json({ error: '`tags` must be an array of strings' });
-    return;
-  }
-  if (typeof properties !== 'object' || properties === null || Array.isArray(properties)) {
-    res.status(400).json({ error: '`properties` must be a plain object' });
-    return;
-  }
-  for (const [k, v] of Object.entries(properties)) {
-    if (typeof k !== 'string' || (typeof v !== 'string' && typeof v !== 'number' && typeof v !== 'boolean')) {
-      res.status(400).json({ error: '`properties` values must be string, number, or boolean' });
-      return;
-    }
-  }
-  const safeDesc: string | undefined = typeof description === 'string' ? description : undefined;
-  const safeId: string | undefined = typeof id === 'string' ? id : undefined;
-
-  // Schema validation
-  const meta = getSpaceMeta(wt.target);
-  const violations = validateEntity(meta ?? {}, { name: name.trim(), type: type.trim(), properties });
-  const validation = applyValidation(meta, violations);
-  if (validation.blocked) {
-    res.status(400).json({ error: 'schema_violation', violations: validation.warnings });
-    return;
-  }
-
-  try {
-    const { entity, warning } = await upsertEntity(wt.target, name.trim(), type.trim(), tags, properties, safeDesc, safeId, undefined, webhookToken(req));
-    const result: Record<string, unknown> = { ...entity };
-    if (warning) result['warning'] = warning;
-    if (validation.warnings.length > 0) result['warnings'] = validation.warnings;
-    res.status(201).json(result);
-  } catch (err) {
-    res.status(500).json({ error: 'Internal server error' });
-  }
-});
-
-// POST /api/brain/spaces/:spaceId/edges — create/upsert an edge
-brainRouter.post('/spaces/:spaceId/edges', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const { from, to, label, weight, type, description, properties, tags } = req.body ?? {};
-  if (!from || typeof from !== 'string') {
-    res.status(400).json({ error: '`from` string required' });
-    return;
-  }
-  if (isStrictLinkage(wt.target) && !UUID_V4_RE.test(from)) {
-    res.status(400).json({ error: '`from` must be a valid UUID v4 (entity ID), not a name' });
-    return;
-  }
-  if (!to || typeof to !== 'string') {
-    res.status(400).json({ error: '`to` string required' });
-    return;
-  }
-  if (isStrictLinkage(wt.target) && !UUID_V4_RE.test(to)) {
-    res.status(400).json({ error: '`to` must be a valid UUID v4 (entity ID), not a name' });
-    return;
-  }
-  if (!label || typeof label !== 'string') {
-    res.status(400).json({ error: '`label` string required' });
-    return;
-  }
-  if (weight !== undefined && typeof weight !== 'number') {
-    res.status(400).json({ error: '`weight` must be a number' });
-    return;
-  }
-  if (type !== undefined && typeof type !== 'string') {
-    res.status(400).json({ error: '`type` must be a string' });
-    return;
-  }
-  if (description !== undefined && typeof description !== 'string') {
-    res.status(400).json({ error: '`description` must be a string' });
-    return;
-  }
-  if (tags !== undefined && (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string'))) {
-    res.status(400).json({ error: '`tags` must be an array of strings' });
-    return;
-  }
-  const safeProps: Record<string, string | number | boolean> | undefined =
-    properties != null && typeof properties === 'object' && !Array.isArray(properties)
-      ? (properties as Record<string, string | number | boolean>)
-      : undefined;
-  const safeTags: string[] | undefined = Array.isArray(tags) ? tags : undefined;
-
-  // Schema validation
-  const meta = getSpaceMeta(wt.target);
-  const violations = validateEdge(meta ?? {}, { label: label.trim(), properties: safeProps });
-  const validation = applyValidation(meta, violations);
-  if (validation.blocked) {
-    res.status(400).json({ error: 'schema_violation', violations: validation.warnings });
-    return;
-  }
-
-  const edge = await upsertEdge(
-    wt.target, from.trim(), to.trim(), label.trim(), weight, type?.trim(),
-    typeof description === 'string' ? description : undefined, safeProps, safeTags,
-    webhookToken(req),
-  );
-  const result: Record<string, unknown> = { ...edge };
-  if (validation.warnings.length > 0) result['warnings'] = validation.warnings;
-  res.status(201).json(result);
-});
-
-// GET /api/brain/spaces/:spaceId/entities
-brainRouter.get('/spaces/:spaceId/entities', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const limit = parseLimit(req.query['limit'], 50, 500);
-  const skip = parseSkip(req.query['skip']);
-  const filter: Record<string, unknown> = {};
-  if (typeof req.query['name'] === 'string') filter['name'] = req.query['name'];
-  if (typeof req.query['type'] === 'string') filter['type'] = req.query['type'];
-  if (typeof req.query['tag'] === 'string') filter['tags'] = req.query['tag'];
-  const all = await collectAcrossMembers(spaceId, mid => listEntities(mid, filter, limit, skip));
-  res.json({ entities: capPage(all, limit), limit, skip });
-});
-
-// GET /api/brain/spaces/:spaceId/entities/by-ids?ids=id1,id2,... — batch fetch up to 100 entities by ID
-brainRouter.get('/spaces/:spaceId/entities/by-ids', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const raw = req.query['ids'];
-  if (typeof raw !== 'string' || !raw.trim()) {
-    res.status(400).json({ error: '`ids` query parameter required (comma-separated)' });
-    return;
-  }
-  const ids = [...new Set(raw.split(',').map(s => s.trim()).filter(Boolean))].slice(0, 100);
-  if (!ids.length) { res.json({ entities: [] }); return; }
-  const all = await collectAcrossMembers(spaceId, mid => listEntities(mid, { _id: { $in: ids } }, 100));
-  res.json({ entities: all });
-});
-
-// GET /api/brain/spaces/:spaceId/entities/by-name?name=... — find entities by name (no type constraint)
-brainRouter.get('/spaces/:spaceId/entities/by-name', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const name = req.query['name'];
-  if (typeof name !== 'string' || !name.trim()) {
-    res.status(400).json({ error: '`name` query parameter required' });
-    return;
-  }
-  // Case-insensitive substring search — escape user input to prevent ReDoS
-  const escaped = escapeRegex(name.trim());
-  const all = await collectAcrossMembers(spaceId, mid => listEntities(mid, { name: { $regex: escaped, $options: 'i' } }, 20));
-  res.json({ entities: all });
-});
-
-// GET /api/brain/spaces/:spaceId/entities/:id
-brainRouter.get('/spaces/:spaceId/entities/:id', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const doc = await findFirstAcrossMembers(spaceId, mid => getEntityById(mid, id));
-  if (doc) { res.json(doc); return; }
-  res.status(404).json({ error: 'Entity not found' });
-});
-
-// DELETE /api/brain/spaces/:spaceId/entities/:id
-brainRouter.delete('/spaces/:spaceId/entities/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const memberIds = resolveMemberSpaces(spaceId);
-  for (const mid of memberIds) {
-    const entity = await getEntityById(mid, id);
-    if (!entity) continue;
-    // Check for inbound references before allowing deletion (only when strictLinkage is on)
-    if (isStrictLinkage(mid)) {
-      const backlinks = await findEntityBacklinks(mid, id);
-      if (backlinks.length > 0) {
-        res.status(409).json({ error: 'Cannot delete: entity has inbound references', backlinks });
-        return;
-      }
-    }
-    if (await deleteEntity(mid, id, webhookToken(req))) {
-      res.status(204).end();
-      return;
-    }
-  }
-  res.status(404).json({ error: 'Entity not found' });
-});
-
-// PATCH /api/brain/spaces/:spaceId/entities/:id — partial update an entity by ID
-brainRouter.patch('/spaces/:spaceId/entities/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const { name, type, description, tags, properties, deleteFields } = req.body ?? {};
-  // Validate deleteFields
-  const dfResult = validateDeleteFields(deleteFields);
-  if (!dfResult.ok) { res.status(400).json({ error: dfResult.error }); return; }
-  const dfPaths: string[] | undefined = Array.isArray(deleteFields) && deleteFields.length > 0 ? deleteFields : undefined;
-  const updates: { name?: string; type?: string; description?: string; tags?: string[]; properties?: Record<string, string | number | boolean> } = {};
-  if (name !== undefined) {
-    if (typeof name !== 'string' || !name.trim()) { res.status(400).json({ error: '`name` must be a non-empty string' }); return; }
-    updates.name = name.trim();
-  }
-  if (type !== undefined) {
-    if (typeof type !== 'string') { res.status(400).json({ error: '`type` must be a string' }); return; }
-    updates.type = type.trim();
-  }
-  if (description !== undefined) {
-    if (typeof description !== 'string') { res.status(400).json({ error: '`description` must be a string' }); return; }
-    updates.description = description;
-  }
-  if (tags !== undefined) {
-    if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) { res.status(400).json({ error: '`tags` must be an array of strings' }); return; }
-    updates.tags = tags;
-  }
-  if (properties !== undefined) {
-    if (typeof properties !== 'object' || properties === null || Array.isArray(properties)) { res.status(400).json({ error: '`properties` must be a plain object' }); return; }
-    updates.properties = properties as Record<string, string | number | boolean>;
-  }
-  if (Object.keys(updates).length === 0 && !dfPaths) { res.status(400).json({ error: 'At least one field must be provided' }); return; }
-  const memberIds = resolveMemberSpaces(wt.target);
-  for (const mid of memberIds) {
-    // Fetch existing entity to validate schema after deleteFields + merge
-    if (dfPaths) {
-      const existing = await getEntityById(mid, id);
-      if (!existing) continue;
-      // Build the resulting entity state to validate against schema
-      const resultName = updates.name ?? existing.name;
-      const resultType = updates.type ?? existing.type;
-      const resultTags = updates.tags !== undefined
-        ? Array.from(new Set([...(existing.tags ?? []), ...updates.tags]))
-        : existing.tags ?? [];
-      const resultProps = updates.properties !== undefined
-        ? { ...(existing.properties ?? {}), ...updates.properties }
-        : { ...(existing.properties ?? {}) };
-      // Build a simulation and apply deleteFields for schema check
-      const sim: Record<string, unknown> = { properties: resultProps, tags: resultTags, description: updates.description !== undefined ? updates.description : existing.description };
-      applyDeleteFieldsPaths(sim, dfPaths);
-      const simProps = (sim['properties'] ?? {}) as Record<string, unknown>;
-      // Schema validation after deleteFields + merge
-      const meta = getSpaceMeta(mid);
-      const violations = validateEntity(meta ?? {}, { name: resultName, type: resultType, properties: simProps, tags: sim['tags'] as string[] });
-      const validation = applyValidation(meta, violations);
-      if (validation.blocked) {
-        res.status(422).json({ error: 'schema_violation', message: 'deleteFields + merge result violates required properties', violations: validation.warnings });
-        return;
-      }
-    }
-    const updated = await updateEntityById(mid, id, updates, dfPaths, webhookToken(req));
-    if (updated) {
-      res.json(updated);
-      return;
-    }
-  }
-  res.status(404).json({ error: 'Entity not found' });
-});
-
-// POST /api/brain/spaces/:spaceId/entities/:survivorId/merge/:absorbedId — merge two entities
-brainRouter.post('/spaces/:spaceId/entities/:survivorId/merge/:absorbedId', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const survivorId = req.params['survivorId'] as string;
-  const absorbedId = req.params['absorbedId'] as string;
-
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-
-  if (!UUID_V4_RE.test(survivorId)) {
-    res.status(400).json({ error: '`survivorId` must be a valid UUID v4' });
-    return;
-  }
-  if (!UUID_V4_RE.test(absorbedId)) {
-    res.status(400).json({ error: '`absorbedId` must be a valid UUID v4' });
-    return;
-  }
-  if (survivorId === absorbedId) {
-    res.status(400).json({ error: 'Cannot merge an entity with itself' });
-    return;
-  }
-
-  if (isProxySpace(spaceId)) {
-    res.status(400).json({ error: 'Entity merge not supported on proxy spaces — target member spaces directly' });
-    return;
-  }
-
-  // Parse resolution map from body (optional)
-  const resolutions: PropertyResolution[] = [];
-  const bodyResolutions = req.body?.resolutions;
-  if (bodyResolutions && Array.isArray(bodyResolutions)) {
-    for (const r of bodyResolutions) {
-      if (typeof r !== 'object' || !r || typeof r.key !== 'string' || typeof r.resolution !== 'string') {
-        res.status(400).json({ error: 'Each resolution must be an object with `key` (string) and `resolution` (string)' });
-        return;
-      }
-      resolutions.push({
-        key: r.key,
-        resolution: r.resolution,
-        ...(r.customValue !== undefined ? { customValue: r.customValue } : {}),
-      });
-    }
-  }
-
-  // Compute merge plan
-  const result = await computeMergePlan(spaceId, survivorId, absorbedId, resolutions);
-  if ('error' in result) {
-    res.status(result.status).json({ error: result.error });
-    return;
-  }
-
-  const { plan, fullyResolved, survivor, absorbed } = result;
-
-  // Validate all provided resolutions
-  for (const conflict of plan.propertyConflicts) {
-    if (!conflict.resolved) continue;
-    const err = validateResolution(conflict.resolution!, conflict.type, conflict.customValue !== undefined);
-    if (err) {
-      res.status(400).json({ error: `Invalid resolution for property '${conflict.key}': ${err}` });
-      return;
-    }
-  }
-
-  // If not fully resolved, return 409 with the plan
-  if (!fullyResolved) {
-    res.status(409).json(plan);
-    return;
-  }
-
-  // All conflicts resolved — execute merge atomically
-  const mergedProperties = applyResolutions(
-    survivor.properties ?? {},
-    absorbed.properties ?? {},
-    plan.propertyConflicts,
-    plan.absorbedOnlyProperties,
-  );
-
-  const mergeResult = await executeMerge(spaceId, survivor, absorbed, mergedProperties, webhookToken(req));
-  const mergedEntity = mergeResult.entity;
-
-  res.json({
-    merged: { ...mergedEntity, embedding: undefined },
-    absorbedId: absorbed._id,
-    relinked: true,
-    duplicateEdgeWarnings: plan.duplicateEdgeWarnings,
-    deletedDuplicateEdgeIds: mergeResult.deletedDuplicateEdgeIds,
-  });
-});
-brainRouter.delete('/spaces/:spaceId/entities', bulkWipeRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  if (isProxySpace(spaceId)) {
-    res.status(400).json({ error: 'Bulk wipe not supported on proxy spaces — target member spaces individually' });
-    return;
-  }
-  if (req.body?.confirm !== true) {
-    res.status(400).json({ error: '`confirm: true` required in request body' });
-    return;
-  }
-  const deleted = await bulkDeleteEntities(spaceId);
-  res.json({ deleted });
-});
-
-// GET /api/brain/spaces/:spaceId/edges
-brainRouter.get('/spaces/:spaceId/edges', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const limit = parseLimit(req.query['limit'], 50, 200);
-  const skip = parseSkip(req.query['skip']);
-  const filter: { from?: string; to?: string; label?: string; type?: string; tag?: string } = {};
-  if (typeof req.query['from'] === 'string') filter.from = req.query['from'];
-  if (typeof req.query['to'] === 'string') filter.to = req.query['to'];
-  if (typeof req.query['label'] === 'string') filter.label = req.query['label'];
-  if (typeof req.query['type'] === 'string') filter.type = req.query['type'];
-  if (typeof req.query['tag'] === 'string') filter.tag = req.query['tag'];
-  const all = await collectAcrossMembers(spaceId, mid => listEdges(mid, filter, limit, skip));
-  // Batch-resolve entity names for from/to so the client can display names instead of raw UUIDs
-  const allEntityIds = [...new Set(all.flatMap(e => [e.from, e.to]))];
-  const nameMap = new Map<string, string>();
-  if (allEntityIds.length) {
-    const nameDocs = await collectAcrossMembers(spaceId, mid =>
-      col<{ _id: string; name: string }>(`${mid}_entities`)
-        .find(asFilter<{ _id: string; name: string }>({ _id: { $in: allEntityIds } }), { projection: { _id: 1, name: 1 } })
-        .toArray());
-    for (const d of nameDocs) nameMap.set(String(d._id), d.name);
-  }
-  const enriched = all.map(e => ({ ...e, fromName: nameMap.get(e.from), toName: nameMap.get(e.to) }));
-  res.json({ edges: capPage(enriched, limit), limit, skip });
-});
-
-// GET /api/brain/spaces/:spaceId/edges/:id
-brainRouter.get('/spaces/:spaceId/edges/:id', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const doc = await findFirstAcrossMembers(spaceId, mid => getEdgeById(mid, id));
-  if (doc) { res.json(doc); return; }
-  res.status(404).json({ error: 'Edge not found' });
-});
-
-// DELETE /api/brain/spaces/:spaceId/edges/:id
-brainRouter.delete('/spaces/:spaceId/edges/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const deleted = await findFirstAcrossMembers(spaceId, mid => deleteEdge(mid, id, webhookToken(req)));
-  if (deleted) { res.status(204).end(); return; }
-  res.status(404).json({ error: 'Edge not found' });
-});
-
-// PATCH /api/brain/spaces/:spaceId/edges/:id — partial update an edge by ID
-brainRouter.patch('/spaces/:spaceId/edges/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const { label, description, tags, properties, weight, type, deleteFields } = req.body ?? {};
-  // Validate deleteFields
-  const dfResult = validateDeleteFields(deleteFields);
-  if (!dfResult.ok) { res.status(400).json({ error: dfResult.error }); return; }
-  const dfPaths: string[] | undefined = Array.isArray(deleteFields) && deleteFields.length > 0 ? deleteFields : undefined;
-  const updates: { label?: string; description?: string; tags?: string[]; properties?: Record<string, string | number | boolean>; weight?: number; type?: string } = {};
-  if (label !== undefined) {
-    if (typeof label !== 'string' || !label.trim()) { res.status(400).json({ error: '`label` must be a non-empty string' }); return; }
-    updates.label = label.trim();
-  }
-  if (description !== undefined) {
-    if (typeof description !== 'string') { res.status(400).json({ error: '`description` must be a string' }); return; }
-    updates.description = description;
-  }
-  if (tags !== undefined) {
-    if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) { res.status(400).json({ error: '`tags` must be an array of strings' }); return; }
-    updates.tags = tags;
-  }
-  if (properties !== undefined) {
-    if (typeof properties !== 'object' || properties === null || Array.isArray(properties)) { res.status(400).json({ error: '`properties` must be a plain object' }); return; }
-    updates.properties = properties as Record<string, string | number | boolean>;
-  }
-  if (weight !== undefined) {
-    if (typeof weight !== 'number') { res.status(400).json({ error: '`weight` must be a number' }); return; }
-    updates.weight = weight;
-  }
-  if (type !== undefined) {
-    if (typeof type !== 'string') { res.status(400).json({ error: '`type` must be a string' }); return; }
-    updates.type = type.trim();
-  }
-  if (Object.keys(updates).length === 0 && !dfPaths) { res.status(400).json({ error: 'At least one field must be provided' }); return; }
-  const memberIds = resolveMemberSpaces(wt.target);
-  for (const mid of memberIds) {
-    // Schema validation after deleteFields + merge
-    if (dfPaths) {
-      const existing = await getEdgeById(mid, id);
-      if (!existing) continue;
-      const resultProps = updates.properties !== undefined
-        ? { ...(existing.properties ?? {}), ...updates.properties }
-        : { ...(existing.properties ?? {}) };
-      const sim: Record<string, unknown> = { properties: resultProps };
-      applyDeleteFieldsPaths(sim, dfPaths);
-      const simProps = (sim['properties'] ?? {}) as Record<string, unknown>;
-      const meta = getSpaceMeta(mid);
-      const violations = validateEdge(meta ?? {}, { label: updates.label ?? existing.label, properties: simProps });
-      const validation = applyValidation(meta, violations);
-      if (validation.blocked) {
-        res.status(422).json({ error: 'schema_violation', message: 'deleteFields + merge result violates required properties', violations: validation.warnings });
-        return;
-      }
-    }
-    const updated = await updateEdgeById(mid, id, updates, dfPaths, webhookToken(req));
-    if (updated) {
-      res.json(updated);
-      return;
-    }
-  }
-  res.status(404).json({ error: 'Edge not found' });
-});
-
-// DELETE /api/brain/spaces/:spaceId/edges — bulk wipe all edges
-brainRouter.delete('/spaces/:spaceId/edges', bulkWipeRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  if (isProxySpace(spaceId)) {
-    res.status(400).json({ error: 'Bulk wipe not supported on proxy spaces — target member spaces individually' });
-    return;
-  }
-  if (req.body?.confirm !== true) {
-    res.status(400).json({ error: '`confirm: true` required in request body' });
-    return;
-  }
-  const deleted = await bulkDeleteEdges(spaceId);
-  res.json({ deleted });
-});
-
-// POST /api/brain/spaces/:spaceId/traverse — graph traversal (BFS)
-brainRouter.post('/spaces/:spaceId/traverse', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const { startId, direction, edgeLabels, maxDepth, limit } = req.body ?? {};
-  if (!startId || typeof startId !== 'string') {
-    res.status(400).json({ error: '`startId` string required' });
-    return;
-  }
-  const validDirections = new Set(['outbound', 'inbound', 'both']);
-  const effectiveDirection: 'outbound' | 'inbound' | 'both' =
-    typeof direction === 'string' && validDirections.has(direction)
-      ? (direction as 'outbound' | 'inbound' | 'both')
-      : 'outbound';
-  const effectiveEdgeLabels: string[] | undefined =
-    Array.isArray(edgeLabels) && edgeLabels.every((l: unknown) => typeof l === 'string')
-      ? edgeLabels
-      : undefined;
-  if (edgeLabels !== undefined && !Array.isArray(edgeLabels)) {
-    res.status(400).json({ error: '`edgeLabels` must be an array of strings' });
-    return;
-  }
-  const rawDepth = typeof maxDepth === 'number' ? maxDepth : 3;
-  const effectiveDepth = Math.min(Math.max(1, rawDepth), 10);
-  const rawLimit = typeof limit === 'number' ? limit : 100;
-  const effectiveLimit = Math.min(Math.max(1, rawLimit), 1000);
-
-  const memberIds = resolveMemberSpaces(spaceId);
-  const result = await traverseGraph(memberIds, startId.trim(), effectiveDirection, effectiveEdgeLabels, effectiveDepth, effectiveLimit);
-  res.json(result);
-});
-
-// ── Chrono CRUD ───────────────────────────────────────────────────────────────
-
-const CHRONO_STATUSES = new Set<ChronoStatus>(['upcoming', 'active', 'completed', 'overdue', 'cancelled']);
-
-// POST /api/brain/spaces/:spaceId/chrono — create a chrono entry
-brainRouter.post('/spaces/:spaceId/chrono', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-
-  const { title, type, startsAt, endsAt, status, confidence, tags, entityIds, memoryIds, description, properties, recurrence } = req.body ?? {};
-  // `recurrence` was previously persisted straight from the request body with NO shape
-  // check — unlike every sibling field — so an arbitrary object could be stored and later
-  // read as a recurrence rule. Shared with MCP so both surfaces enforce the same shape.
-  const recCheck = parseRecurrence(recurrence);
-  if (!recCheck.ok) { res.status(400).json({ error: recCheck.error }); return; }
-  const safeRecurrence = recCheck.value;
-  if (!title || typeof title !== 'string') {
-    res.status(400).json({ error: '`title` string required' }); return;
-  }
-  const meta = getSpaceMeta(wt.target);
-  const allowedChronoTypes = getAllowedChronoTypes(meta);
-  if (!type || !allowedChronoTypes.has(type)) {
-    res.status(400).json({ error: `\`type\` must be one of: ${[...allowedChronoTypes].join(', ')}` }); return;
-  }
-  if (!startsAt || typeof startsAt !== 'string') {
-    res.status(400).json({ error: '`startsAt` ISO8601 string required' }); return;
-  }
-  if (endsAt !== undefined && typeof endsAt !== 'string') {
-    res.status(400).json({ error: '`endsAt` must be an ISO8601 string' }); return;
-  }
-  if (status !== undefined && !CHRONO_STATUSES.has(status)) {
-    res.status(400).json({ error: '`status` must be one of: upcoming, active, completed, overdue, cancelled' }); return;
-  }
-  if (confidence !== undefined && (typeof confidence !== 'number' || confidence < 0 || confidence > 1)) {
-    res.status(400).json({ error: '`confidence` must be a number between 0 and 1' }); return;
-  }
-  if (tags !== undefined && (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string'))) {
-    res.status(400).json({ error: '`tags` must be an array of strings' }); return;
-  }
-  if (entityIds !== undefined && (!Array.isArray(entityIds) || entityIds.some((t: unknown) => typeof t !== 'string'))) {
-    res.status(400).json({ error: '`entityIds` must be an array of strings' }); return;
-  }
-  if (entityIds !== undefined) {
-    if (isStrictLinkage(wt.target)) {
-      const invalidEIds = (entityIds as string[]).filter((id: string) => !UUID_V4_RE.test(id));
-      if (invalidEIds.length > 0) { res.status(400).json({ error: '`entityIds` must contain valid UUID v4 values (entity IDs), not names', invalid: invalidEIds }); return; }
-    }
-  }
-  if (memoryIds !== undefined && (!Array.isArray(memoryIds) || memoryIds.some((t: unknown) => typeof t !== 'string'))) {
-    res.status(400).json({ error: '`memoryIds` must be an array of strings' }); return;
-  }
-  if (memoryIds !== undefined) {
-    if (isStrictLinkage(wt.target)) {
-      const invalidMIds = (memoryIds as string[]).filter((id: string) => !UUID_V4_RE.test(id));
-      if (invalidMIds.length > 0) { res.status(400).json({ error: '`memoryIds` must contain valid UUID v4 values (memory IDs), not names', invalid: invalidMIds }); return; }
-    }
-  }
-  if (description !== undefined && typeof description !== 'string') {
-    res.status(400).json({ error: '`description` must be a string' }); return;
-  }
-  if (properties !== undefined && (typeof properties !== 'object' || properties === null || Array.isArray(properties))) {
-    res.status(400).json({ error: '`properties` must be a plain object' }); return;
-  }
-  const safeProps: Record<string, string | number | boolean> | undefined =
-    properties != null && typeof properties === 'object' && !Array.isArray(properties)
-      ? (properties as Record<string, string | number | boolean>)
-      : undefined;
-
-  // Schema validation (meta already resolved above)
-  const violations = validateChrono(meta ?? {}, { type, properties: safeProps });
-  const validation = applyValidation(meta, violations);
-  if (validation.blocked) {
-    res.status(400).json({ error: 'schema_violation', violations: validation.warnings });
-    return;
-  }
-
-  const entry = await createChrono(wt.target, {
-    title: title.trim(), type, startsAt, endsAt, status, confidence,
-    tags, entityIds, memoryIds, description, properties: safeProps, recurrence: safeRecurrence,
-  }, webhookToken(req));
-  const result: Record<string, unknown> = { ...entry };
-  if (validation.warnings.length > 0) result['warnings'] = validation.warnings;
-  res.status(201).json(result);
-});
-
-// POST /api/brain/spaces/:spaceId/chrono/:id — update a chrono entry
-brainRouter.post('/spaces/:spaceId/chrono/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-
-  const { title, type, startsAt, endsAt, status, confidence, tags, entityIds, memoryIds, description, properties, recurrence } = req.body ?? {};
-  // `recurrence` was previously persisted straight from the request body with NO shape
-  // check — unlike every sibling field — so an arbitrary object could be stored and later
-  // read as a recurrence rule. Shared with MCP so both surfaces enforce the same shape.
-  const recCheck = parseRecurrence(recurrence);
-  if (!recCheck.ok) { res.status(400).json({ error: recCheck.error }); return; }
-  const safeRecurrence = recCheck.value;
-  if (status !== undefined && !CHRONO_STATUSES.has(status)) {
-    res.status(400).json({ error: '`status` must be one of: upcoming, active, completed, overdue, cancelled' }); return;
-  }
-  if (type !== undefined) {
-    const allowedChronoTypes = getAllowedChronoTypes(getSpaceMeta(wt.target));
-    if (!allowedChronoTypes.has(type)) {
-      res.status(400).json({ error: `\`type\` must be one of: ${[...allowedChronoTypes].join(', ')}` }); return;
-    }
-  }
-  if (confidence !== undefined && (typeof confidence !== 'number' || confidence < 0 || confidence > 1)) {
-    res.status(400).json({ error: '`confidence` must be a number between 0 and 1' }); return;
-  }
-  if (entityIds !== undefined && Array.isArray(entityIds) && isStrictLinkage(wt.target)) {
-    const invalidEIds = entityIds.filter((id: string) => !UUID_V4_RE.test(id));
-    if (invalidEIds.length > 0) { res.status(400).json({ error: '`entityIds` must contain valid UUID v4 values (entity IDs), not names', invalid: invalidEIds }); return; }
-  }
-  if (memoryIds !== undefined && Array.isArray(memoryIds) && isStrictLinkage(wt.target)) {
-    const invalidMIds = memoryIds.filter((id: string) => !UUID_V4_RE.test(id));
-    if (invalidMIds.length > 0) { res.status(400).json({ error: '`memoryIds` must contain valid UUID v4 values (memory IDs), not names', invalid: invalidMIds }); return; }
-  }
-  if (properties !== undefined && (typeof properties !== 'object' || properties === null || Array.isArray(properties))) {
-    res.status(400).json({ error: '`properties` must be a plain object' }); return;
-  }
-  const safeProps: Record<string, string | number | boolean> | undefined =
-    properties != null && typeof properties === 'object' && !Array.isArray(properties)
-      ? (properties as Record<string, string | number | boolean>)
-      : undefined;
-
-  const updated = await updateChrono(wt.target, id, {
-    title, type, startsAt, endsAt, status, confidence,
-    tags, entityIds, memoryIds, description, properties: safeProps, recurrence: safeRecurrence,
-  }, webhookToken(req));
-  if (!updated) { res.status(404).json({ error: 'Chrono entry not found' }); return; }
-  res.json(updated);
-});
-
-// PATCH /api/brain/spaces/:spaceId/chrono/:id — partial update a chrono entry by ID
-brainRouter.patch('/spaces/:spaceId/chrono/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-
-  const { title, type, startsAt, endsAt, status, confidence, tags, entityIds, memoryIds, description, properties, recurrence } = req.body ?? {};
-  // `recurrence` was previously persisted straight from the request body with NO shape
-  // check — unlike every sibling field — so an arbitrary object could be stored and later
-  // read as a recurrence rule. Shared with MCP so both surfaces enforce the same shape.
-  const recCheck = parseRecurrence(recurrence);
-  if (!recCheck.ok) { res.status(400).json({ error: recCheck.error }); return; }
-  const safeRecurrence = recCheck.value;
-  if (status !== undefined && !CHRONO_STATUSES.has(status)) {
-    res.status(400).json({ error: '`status` must be one of: upcoming, active, completed, overdue, cancelled' }); return;
-  }
-  if (type !== undefined) {
-    const allowedChronoTypes = getAllowedChronoTypes(getSpaceMeta(wt.target));
-    if (!allowedChronoTypes.has(type)) {
-      res.status(400).json({ error: `\`type\` must be one of: ${[...allowedChronoTypes].join(', ')}` }); return;
-    }
-  }
-  if (confidence !== undefined && (typeof confidence !== 'number' || confidence < 0 || confidence > 1)) {
-    res.status(400).json({ error: '`confidence` must be a number between 0 and 1' }); return;
-  }
-  if (entityIds !== undefined && Array.isArray(entityIds) && isStrictLinkage(wt.target)) {
-    const invalidEIds = entityIds.filter((id: string) => !UUID_V4_RE.test(id));
-    if (invalidEIds.length > 0) { res.status(400).json({ error: '`entityIds` must contain valid UUID v4 values (entity IDs), not names', invalid: invalidEIds }); return; }
-  }
-  if (memoryIds !== undefined && Array.isArray(memoryIds) && isStrictLinkage(wt.target)) {
-    const invalidMIds = memoryIds.filter((id: string) => !UUID_V4_RE.test(id));
-    if (invalidMIds.length > 0) { res.status(400).json({ error: '`memoryIds` must contain valid UUID v4 values (memory IDs), not names', invalid: invalidMIds }); return; }
-  }
-  if (properties !== undefined && (typeof properties !== 'object' || properties === null || Array.isArray(properties))) {
-    res.status(400).json({ error: '`properties` must be a plain object' }); return;
-  }
-  const safeProps: Record<string, string | number | boolean> | undefined =
-    properties != null && typeof properties === 'object' && !Array.isArray(properties)
-      ? (properties as Record<string, string | number | boolean>)
-      : undefined;
-
-  const updated = await findFirstAcrossMembers(wt.target, mid => updateChrono(mid, id, {
-    title, type, startsAt, endsAt, status, confidence,
-    tags, entityIds, memoryIds, description, properties: safeProps, recurrence: safeRecurrence,
-  }, webhookToken(req)));
-  if (updated) { res.json(updated); return; }
-  res.status(404).json({ error: 'Chrono entry not found' });
-});
-
-// GET /api/brain/spaces/:spaceId/chrono/:id
-brainRouter.get('/spaces/:spaceId/chrono/:id', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const doc = await findFirstAcrossMembers(spaceId, mid => getChronoById(mid, id));
-  if (doc) { res.json(doc); return; }
-  res.status(404).json({ error: 'Chrono entry not found' });
-});
-
-// GET /api/brain/spaces/:spaceId/chrono
-brainRouter.get('/spaces/:spaceId/chrono', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const limit = parseLimit(req.query['limit'], 50, 500);
-  const skip = parseSkip(req.query['skip']);
-  const filter: ChronoFilter = {};
-  if (typeof req.query['status'] === 'string') filter.status = req.query['status'];
-  if (typeof req.query['type'] === 'string') filter.type = req.query['type'];
-
-  // tags — comma-separated or repeated — AND semantics
-  if (Array.isArray(req.query['tags'])) {
-    filter.tags = (req.query['tags'] as string[]).flatMap(t => t.split(',').map(s => s.trim())).filter(Boolean);
-  } else if (typeof req.query['tags'] === 'string') {
-    filter.tags = req.query['tags'].split(',').map(s => s.trim()).filter(Boolean);
-  } else if (typeof req.query['tag'] === 'string') {
-    filter.tags = [req.query['tag']];
-  }
-
-  // tagsAny — comma-separated or repeated — OR semantics
-  if (Array.isArray(req.query['tagsAny'])) {
-    filter.tagsAny = (req.query['tagsAny'] as string[]).flatMap(t => t.split(',').map(s => s.trim())).filter(Boolean);
-  } else if (typeof req.query['tagsAny'] === 'string') {
-    filter.tagsAny = req.query['tagsAny'].split(',').map(s => s.trim()).filter(Boolean);
-  }
-
-  if (typeof req.query['after'] === 'string') filter.after = req.query['after'];
-  if (typeof req.query['before'] === 'string') filter.before = req.query['before'];
-  if (typeof req.query['search'] === 'string') filter.search = req.query['search'];
-
-  const all = await collectAcrossMembers(spaceId, mid => listChrono(mid, filter, limit, skip));
-  res.json({ chrono: capPage(all, limit), limit, skip });
-});
-
-// DELETE /api/brain/spaces/:spaceId/chrono/:id
-brainRouter.delete('/spaces/:spaceId/chrono/:id', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const id = req.params['id'] as string;
-  const deleted = await findFirstAcrossMembers(spaceId, mid => deleteChrono(mid, id, webhookToken(req)));
-  if (deleted) { res.status(204).end(); return; }
-  res.status(404).json({ error: 'Chrono entry not found' });
-});
-
-// DELETE /api/brain/spaces/:spaceId/chrono — bulk wipe all chrono entries
-brainRouter.delete('/spaces/:spaceId/chrono', bulkWipeRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  if (isProxySpace(spaceId)) {
-    res.status(400).json({ error: 'Bulk wipe not supported on proxy spaces — target member spaces individually' });
-    return;
-  }
-  if (req.body?.confirm !== true) {
-    res.status(400).json({ error: '`confirm: true` required in request body' });
-    return;
-  }
-  const deleted = await bulkDeleteChrono(spaceId);
-  res.json({ deleted });
-});
-
-// GET /api/brain/spaces/:spaceId/files — list file metadata records
-brainRouter.get('/spaces/:spaceId/files', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const limit = parseLimit(req.query['limit'], 50, 200);
-  const skip = parseSkip(req.query['skip']);
-  // By default exclude chunk records (parentFileId set) so the file manager only shows
-  // top-level files. Pass ?includeChunks=true to see all records (e.g. for debugging).
-  const includeChunks = req.query['includeChunks'] === 'true';
-  const filter: Record<string, unknown> = {};
-  if (!includeChunks) filter['parentFileId'] = { $exists: false };
-  if (typeof req.query['tag'] === 'string') filter['tags'] = req.query['tag'];
-  if (typeof req.query['path'] === 'string') filter['path'] = toDocId(req.query['path']);
-  const all = await collectAcrossMembers(spaceId, mid =>
-    col(`${mid}_files`)
-      .find(asFilter(filter))
-      .sort({ updatedAt: -1 })
-      .skip(skip)
-      .limit(limit)
-      .toArray(),
-  );
-  res.json({ files: capPage(all, limit), limit, skip });
-});
-
-// DELETE /api/brain/spaces/:spaceId/files — delete file metadata record by path (does NOT delete the file on disk)
-brainRouter.delete('/spaces/:spaceId/files', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` }); return;
-  }
-  const path = req.query['path'];
-  if (typeof path !== 'string' || !path.trim()) {
-    res.status(400).json({ error: '`path` query parameter required' }); return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const memberIds = resolveMemberSpaces(wt.target);
-  const norm = toDocId(path);
-  // Guard: a metadata record may only be removed if its file is gone (orphan) or the
-  // record is flagged deleted. Deleting the metadata of a file that still exists would
-  // silently orphan a live file — refuse and tell the caller to delete the file itself.
-  for (const mid of memberIds) {
-    const rec = await col<FileMetaDoc>(`${mid}_files`).findOne(asFilter<FileMetaDoc>({ _id: norm })) as FileMetaDoc | null;
-    if (rec && !rec.deletedAt && await fileExists(mid, norm)) {
-      res.status(409).json({
-        error: 'Cannot delete metadata while the file still exists. Delete the file itself (which also removes its metadata), or enable softDeleteFileMeta and delete the file first.',
-      });
-      return;
-    }
-  }
-  for (const mid of memberIds) {
-    await deleteFileMeta(mid, path);
-  }
-  res.status(204).end();
-});
-
-// PATCH /api/brain/spaces/:spaceId/files — update file metadata by path (query param ?path=)
-brainRouter.patch('/spaces/:spaceId/files', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` }); return;
-  }
-  const path = req.query['path'];
-  if (typeof path !== 'string' || !path.trim()) {
-    res.status(400).json({ error: '`path` query parameter required' }); return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-
-  const { description, tags, entityIds, chronoIds, memoryIds, properties } = req.body ?? {};
-  if (tags !== undefined && !Array.isArray(tags)) { res.status(400).json({ error: '`tags` must be an array' }); return; }
-  if (entityIds !== undefined && !Array.isArray(entityIds)) { res.status(400).json({ error: '`entityIds` must be an array' }); return; }
-  if (chronoIds !== undefined && !Array.isArray(chronoIds)) { res.status(400).json({ error: '`chronoIds` must be an array' }); return; }
-  if (memoryIds !== undefined && !Array.isArray(memoryIds)) { res.status(400).json({ error: '`memoryIds` must be an array' }); return; }
-  if (properties !== undefined && (typeof properties !== 'object' || properties === null || Array.isArray(properties))) {
-    res.status(400).json({ error: '`properties` must be a plain object' }); return;
-  }
-
-  const updated = await findFirstAcrossMembers(wt.target,
-    mid => updateFileMeta(mid, path, { description, tags, entityIds, chronoIds, memoryIds, properties }));
-  if (updated) { res.json(updated); return; }
-  res.status(404).json({ error: 'File metadata record not found' });
-});
-
-// POST /api/brain/spaces/:spaceId/query — structured query with filter/projection
-brainRouter.post('/spaces/:spaceId/query', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const { collection, filter, projection, limit, maxTimeMS } = req.body ?? {};
-  const validCollections = ['memories', 'entities', 'edges', 'chrono', 'files'] as const;
-  if (!validCollections.includes(collection)) {
-    res.status(400).json({ error: `collection must be one of: ${validCollections.join(', ')}` });
-    return;
-  }
-  const safeFilter: Record<string, unknown> =
-    filter != null && typeof filter === 'object' && !Array.isArray(filter)
-      ? (filter as Record<string, unknown>)
-      : {};
-  const safeProjection: Record<string, unknown> | undefined =
-    projection != null && typeof projection === 'object' && !Array.isArray(projection)
-      ? (projection as Record<string, unknown>)
-      : undefined;
-  const safeLimit = typeof limit === 'number' ? limit : 20;
-  const safeMaxTimeMS = typeof maxTimeMS === 'number' ? maxTimeMS : 5000;
-
-  try {
-    const docs = await collectAcrossMembers(spaceId, mid =>
-      queryBrain(
-        mid,
-        collection as typeof validCollections[number],
-        safeFilter,
-        safeProjection,
-        safeLimit,
-        safeMaxTimeMS,
-      ),
-    );
-    res.json({ results: docs, collection, count: docs.length });
-  } catch (err: unknown) {
-    const msg = err instanceof Error ? err.message : String(err);
-    res.status(400).json({ error: msg });
-  }
-});
-
-/** One entry in a graph-augmented recall response (traverse > 0). */
-interface RecallTraverseItem {
-  /** Vector similarity score for seeds; null for traversal-reached records. */
-  score: number | null;
-  source: 'recall' | 'traverse';
-  /** 0 = seed, 1 = one edge away, etc. */
-  hops: number;
-  /** Edge chain connecting this record to its seed (empty for seeds). */
-  path: { from: string; label: string; to: string }[];
-  spaceId: string;
-  type: string;
-  record: unknown;
-}
-
-// POST /api/brain/spaces/:spaceId/recall — semantic vector search by natural language query
-brainRouter.post('/spaces/:spaceId/recall', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const { query, topK, types, minScore, filter, traverse, tags, minPerType } = req.body ?? {};
-  if (!query || typeof query !== 'string' || !query.trim()) {
-    res.status(400).json({ error: 'query must be a non-empty string' });
-    return;
-  }
-  const safeTopK = typeof topK === 'number' ? Math.min(Math.max(topK, 1), 100) : 10;
-  const safeTypes = Array.isArray(types) ? types.filter((t: unknown): t is RecallKnowledgeType => typeof t === 'string') : undefined;
-  const safeMinScore = typeof minScore === 'number' ? minScore : undefined;
-
-  // `tags` and `minPerType` are supported by recall() but were previously hardcoded to
-  // undefined here, so they were reachable only via MCP / the internal function.
-  let safeTags: string[] | undefined;
-  if (tags != null) {
-    if (!Array.isArray(tags) || tags.some((t: unknown) => typeof t !== 'string')) {
-      res.status(400).json({ error: 'tags must be an array of strings' });
-      return;
-    }
-    safeTags = (tags as string[]).filter(t => t.trim().length > 0);
-    if (safeTags.length === 0) safeTags = undefined;
-  }
-
-  // Per-type minimums: guarantee at least N hits of a given knowledge type. Each value
-  // is clamped to [0, topK] — asking for more of a type than the total result size is
-  // meaningless, and an unbounded value would widen the underlying per-type searches.
-  let safeMinPerType: Partial<Record<RecallKnowledgeType, number>> | undefined;
-  if (minPerType != null) {
-    if (typeof minPerType !== 'object' || Array.isArray(minPerType)) {
-      res.status(400).json({ error: 'minPerType must be an object mapping knowledge type -> minimum count' });
-      return;
-    }
-    const acc: Partial<Record<RecallKnowledgeType, number>> = {};
-    for (const [key, raw] of Object.entries(minPerType as Record<string, unknown>)) {
-      if (typeof raw !== 'number' || !Number.isInteger(raw) || raw < 0) {
-        res.status(400).json({ error: `minPerType.${key} must be a non-negative integer` });
-        return;
-      }
-      acc[key as RecallKnowledgeType] = Math.min(raw, safeTopK);
-    }
-    if (Object.keys(acc).length > 0) safeMinPerType = acc;
-  }
-
-  // Graph-traversal expansion depth. 0 (default) = classic recall, unchanged.
-  // Rejected rather than clamped past the cap so an obviously-wrong depth surfaces.
-  let safeTraverse = 0;
-  if (traverse != null) {
-    if (typeof traverse !== 'number' || !Number.isInteger(traverse) || traverse < 0 || traverse > MAX_RECALL_TRAVERSE) {
-      res.status(400).json({ error: `traverse must be an integer between 0 and ${MAX_RECALL_TRAVERSE}` });
-      return;
-    }
-    safeTraverse = traverse;
-  }
-
-  let safeFilter: FilterExpression | undefined;
-  if (filter != null) {
-    if (typeof filter !== 'object' || Array.isArray(filter)) {
-      res.status(400).json({ error: 'filter must be an object' });
-      return;
-    }
-    const filterErr = validateFilterExpression(filter as FilterExpression);
-    if (filterErr) {
-      res.status(400).json({ error: filterErr });
-      return;
-    }
-    safeFilter = filter as FilterExpression;
-  }
-
-  try {
-    const memberIds = resolveMemberSpaces(spaceId);
-    const all = (await Promise.all(
-      memberIds.map(mid => recall(mid, query.trim(), safeTopK, safeTags, safeTypes, safeMinPerType, safeMinScore, safeFilter)),
-    )).flat();
-    all.sort((x, y) => (y.score ?? 0) - (x.score ?? 0));
-    const seeds = all.slice(0, safeTopK);
-
-    if (safeTraverse === 0) {
-      res.json({ results: seeds, count: seeds.length });
-      return;
-    }
-
-    // Graph-augmented recall: expand seeds along edges, cap the combined output.
-    const totalCap = safeTopK * (safeTraverse + 1) * 4;
-    const neighbours = await traverseRecallSeeds(
-      memberIds,
-      seeds.map(s => ({ _id: s._id, spaceId: s.spaceId })),
-      safeTraverse,
-      Math.max(0, totalCap - seeds.length),
-    );
-    const results: RecallTraverseItem[] = [
-      ...seeds.map(s => ({ score: s.score, source: 'recall' as const, hops: 0, path: [], spaceId: s.spaceId, type: s.type, record: s })),
-      ...neighbours.map(n => ({ score: null, source: 'traverse' as const, hops: n.hops, path: n.path, spaceId: n.spaceId, type: 'entity' as const, record: n.record })),
-    ];
-    res.json({ results, count: results.length, traverseDepth: safeTraverse });
-  } catch (err: unknown) {
-    const msg = err instanceof Error ? err.message : String(err);
-    res.status(400).json({ error: msg });
-  }
-});
-
-// POST /api/brain/spaces/:spaceId/find-similar — vector similarity search by existing entry ID
-const VALID_ENTRY_TYPES = new Set(['memory', 'entity', 'edge', 'chrono', 'file']);
-
-brainRouter.post('/spaces/:spaceId/find-similar', globalRateLimit, requireSpaceAuth, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-
-  const body = (req.body ?? {}) as Record<string, unknown>;
-  const entryId = typeof body['entryId'] === 'string' ? body['entryId'].trim() : '';
-  const entryType = typeof body['entryType'] === 'string' ? body['entryType'].trim() : '';
-  const topK = typeof body['topK'] === 'number' ? Math.min(Math.max(body['topK'], 1), 100) : 10;
-  const minScore = typeof body['minScore'] === 'number' ? body['minScore'] : undefined;
-  const crossSpace = body['crossSpace'] === true;
-  const targetTypes = Array.isArray(body['targetTypes'])
-    ? (body['targetTypes'] as unknown[]).filter((t): t is RecallKnowledgeType => typeof t === 'string' && VALID_ENTRY_TYPES.has(t))
-    : undefined;
-
-  if (!entryId || !UUID_V4_RE.test(entryId)) {
-    res.status(400).json({ error: 'entryId must be a valid UUID v4' });
-    return;
-  }
-  if (!VALID_ENTRY_TYPES.has(entryType)) {
-    res.status(400).json({ error: `entryType must be one of: ${[...VALID_ENTRY_TYPES].join(', ')}` });
-    return;
-  }
-
-  // Determine cross-space search scope
-  let crossSpaceIds: string[] | undefined;
-  if (crossSpace) {
-    const tokenSpaces = req.authToken?.spaces;
-    crossSpaceIds = cfg.spaces
-      .filter(s => !tokenSpaces || tokenSpaces.includes(s.id))
-      .map(s => s.id);
-  }
-
-  try {
-    const result = await findSimilar(
-      spaceId,
-      entryId,
-      entryType as RecallKnowledgeType,
-      topK,
-      targetTypes,
-      minScore,
-      crossSpaceIds,
-    );
-    res.json(result);
-  } catch (err: unknown) {
-    if (err instanceof NotFoundError) {
-      res.status(404).json({ error: err.message });
-    } else {
-      const msg = err instanceof Error ? err.message : String(err);
-      res.status(400).json({ error: msg });
-    }
-  }
-});
-
-brainRouter.get('/spaces/:spaceId/reindex-status', globalRateLimit, requireSpaceAuth, (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const memberIds = resolveMemberSpaces(spaceId);
-  const needs = memberIds.some(mid => needsReindex(mid));
-  res.json({ spaceId, needsReindex: needs });
-});
-
-// POST /api/brain/spaces/:spaceId/reindex
-// Re-embeds all memories in a space using the currently configured model.
-// Long-running: may take minutes for large spaces. Progress is logged server-side.
-brainRouter.post('/spaces/:spaceId/reindex', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-
-  if (reindexJobRunning) {
-    res.status(409).json({ error: 'Reindex already in progress' });
-    return;
-  }
-
-  const memberIds = resolveMemberSpaces(spaceId);
-  reindexJobRunning = true;
-  reindexInProgress.set(1);
-  res.json({ spaceId, reindexed: 0, errors: 0, status: 'started' });
-
-  // Start heavy work on the next turn so HTTP headers flush immediately.
-  setImmediate(() => {
-    void (async () => {
-      let reindexed = 0;
-      let errors = 0;
-      try {
-        for (const mid of memberIds) {
-        const BATCH = 50;
-
-        // Re-embed memories
-        {
-          let cursor: string | null = null;
-          // eslint-disable-next-line no-constant-condition
-          while (true) {
-            const q: Record<string, unknown> = cursor ? { _id: { $gt: cursor } } : {};
-            const batch: MemoryDoc[] = await col<MemoryDoc>(`${mid}_memories`)
-              .find(asFilter<MemoryDoc>(q), { projection: { _id: 1, fact: 1, tags: 1, entityIds: 1, description: 1, properties: 1 } })
-              .sort({ _id: 1 })
-              .limit(BATCH)
-              .toArray() as MemoryDoc[];
-            if (batch.length === 0) break;
-            for (const doc of batch) {
-              try {
-                const entityIds: string[] = Array.isArray(doc.entityIds) ? doc.entityIds : [];
-                const entityDocs = entityIds.length > 0
-                  ? await col<EntityDoc>(`${mid}_entities`)
-                      .find(asFilter<EntityDoc>({ _id: { $in: entityIds } }), { projection: { name: 1 } })
-                      .toArray() as Array<{ name: string }>
-                  : [];
-                const entityNames = entityDocs.map(e => e.name);
-                const result = await embed(memoryEmbedText(doc.fact, doc.tags ?? [], entityNames, doc.description, doc.properties));
-                await col<MemoryDoc>(`${mid}_memories`).updateOne(
-                  { _id: doc._id },
-                  { $set: { embedding: result.vector, embeddingModel: result.model } },
-                );
-                reindexed++;
-              } catch { errors++; }
-            }
-            cursor = batch[batch.length - 1]?._id ?? null;
-          }
-        }
-
-        // Re-embed entities (name + type + tags + description + properties)
-        {
-          let cursor: string | null = null;
-          // eslint-disable-next-line no-constant-condition
-          while (true) {
-            const q: Record<string, unknown> = cursor ? { _id: { $gt: cursor } } : {};
-            const batch: EntityDoc[] = await col<EntityDoc>(`${mid}_entities`)
-              .find(asFilter<EntityDoc>(q), { projection: { _id: 1, name: 1, type: 1, tags: 1, description: 1, properties: 1 } })
-              .sort({ _id: 1 })
-              .limit(BATCH)
-              .toArray() as EntityDoc[];
-            if (batch.length === 0) break;
-            for (const doc of batch) {
-              try {
-                const result = await embed(entityEmbedText(doc.name, doc.type, doc.tags ?? [], doc.description, doc.properties ?? {}));
-                await col<EntityDoc>(`${mid}_entities`).updateOne(
-                  { _id: doc._id },
-                  { $set: { embedding: result.vector, embeddingModel: result.model } },
-                );
-                reindexed++;
-              } catch { errors++; }
-            }
-            cursor = batch[batch.length - 1]?._id ?? null;
-          }
-        }
-
-        // Re-embed edges (tags + from-name + label + to-name + type + description + properties)
-        {
-          let cursor: string | null = null;
-          // eslint-disable-next-line no-constant-condition
-          while (true) {
-            const q: Record<string, unknown> = cursor ? { _id: { $gt: cursor } } : {};
-            const batch: EdgeDoc[] = await col<EdgeDoc>(`${mid}_edges`)
-              .find(asFilter<EdgeDoc>(q), { projection: { _id: 1, from: 1, label: 1, to: 1, type: 1, tags: 1, description: 1, properties: 1 } })
-              .sort({ _id: 1 })
-              .limit(BATCH)
-              .toArray() as EdgeDoc[];
-            if (batch.length === 0) break;
-            for (const doc of batch) {
-              try {
-                // Resolve from/to to entity NAMES (not IDs) and include properties — matching
-                // edgeEmbedText so a reindex reproduces exactly what upsertEdge embedded.
-                const [fromName, toName] = await resolveEdgeEntityNames(mid, doc.from, doc.to);
-                const result = await embed(edgeEmbedText(fromName, doc.label, toName, doc.tags ?? [], doc.type, doc.description, doc.properties));
-                await col<EdgeDoc>(`${mid}_edges`).updateOne(
-                  { _id: doc._id },
-                  { $set: { embedding: result.vector, embeddingModel: result.model } },
-                );
-                reindexed++;
-              } catch { errors++; }
-            }
-            cursor = batch[batch.length - 1]?._id ?? null;
-          }
-        }
-
-        // Re-embed chrono (type + status + title + tags + description + properties)
-        {
-          let cursor: string | null = null;
-          // eslint-disable-next-line no-constant-condition
-          while (true) {
-            const q: Record<string, unknown> = cursor ? { _id: { $gt: cursor } } : {};
-            const batch: ChronoEntry[] = await col<ChronoEntry>(`${mid}_chrono`)
-              .find(asFilter<ChronoEntry>(q), { projection: { _id: 1, title: 1, type: 1, status: 1, description: 1, tags: 1, properties: 1 } })
-              .sort({ _id: 1 })
-              .limit(BATCH)
-              .toArray() as ChronoEntry[];
-            if (batch.length === 0) break;
-            for (const doc of batch) {
-              try {
-                const result = await embed(chronoEmbedText(doc.title, doc.type, doc.status, doc.description, doc.tags ?? [], doc.properties));
-                await col<ChronoEntry>(`${mid}_chrono`).updateOne(
-                  { _id: doc._id },
-                  { $set: { embedding: result.vector, embeddingModel: result.model } },
-                );
-                reindexed++;
-              } catch { errors++; }
-            }
-            cursor = batch[batch.length - 1]?._id ?? null;
-          }
-        }
-
-        // Re-embed files (path + entity names + tags + description + property values)
-        {
-          let cursor: string | null = null;
-          // eslint-disable-next-line no-constant-condition
-          while (true) {
-            // Exclude chunk records (parentFileId set) — they have their own embedding logic
-            const q: Record<string, unknown> = cursor
-              ? { _id: { $gt: cursor }, parentFileId: { $exists: false } }
-              : { parentFileId: { $exists: false } };
-            const batch: FileMetaDoc[] = await col<FileMetaDoc>(`${mid}_files`)
-              .find(asFilter<FileMetaDoc>(q), { projection: { _id: 1, path: 1, tags: 1, description: 1, properties: 1, entityIds: 1 } })
-              .sort({ _id: 1 })
-              .limit(BATCH)
-              .toArray() as FileMetaDoc[];
-            if (batch.length === 0) break;
-            for (const doc of batch) {
-              try {
-                const entityIds: string[] = Array.isArray(doc.entityIds) ? doc.entityIds : [];
-                const entityDocs = entityIds.length > 0
-                  ? await col<EntityDoc>(`${mid}_entities`)
-                      .find(asFilter<EntityDoc>({ _id: { $in: entityIds } }), { projection: { name: 1 } })
-                      .toArray() as Array<{ name: string }>
-                  : [];
-                const entityNames = entityDocs.map(e => e.name);
-                const result = await embed(fileEmbedText(doc.path, doc.tags ?? [], doc.description, doc.properties, entityNames));
-                await col<FileMetaDoc>(`${mid}_files`).updateOne(
-                  { _id: doc._id },
-                  { $set: { embedding: result.vector, embeddingModel: result.model } },
-                );
-                reindexed++;
-              } catch { errors++; }
-            }
-            cursor = batch[batch.length - 1]?._id ?? null;
-          }
-        }
-
-          clearReindexFlag(mid);
-        }
-        log.info(`Reindex completed for space '${spaceId}': reindexed=${reindexed}, errors=${errors}`);
-      } catch (err) {
-        log.error(`Reindex job failed for space '${spaceId}': ${String(err)}`);
-      } finally {
-        reindexJobRunning = false;
-        reindexInProgress.set(0);
-      }
-    })();
-  });
-});
-
-// ── Bulk write ────────────────────────────────────────────────────────────────
-
-/**
- * POST /api/brain/spaces/:spaceId/bulk
- *
- * Batch upsert memories, entities, edges, and chrono entries in a single
- * request.  Processing order: memories → entities → edges → chrono, so edges
- * referencing newly created entities within the same batch resolve correctly.
- *
- * All four arrays are optional.  Entries that fail per-item validation are
- * recorded in `errors` and do not abort the remaining batch items.
- */
-brainRouter.post('/spaces/:spaceId/bulk', globalRateLimit, requireSpaceAuth, denyReadOnly, async (req, res) => {
-  const spaceId = req.params['spaceId'] as string;
-  const cfg = getConfig();
-  if (!cfg.spaces.some(s => s.id === spaceId)) {
-    res.status(404).json({ error: `Space '${spaceId}' not found` });
-    return;
-  }
-  const wt = resolveWriteTarget(spaceId, req.query['targetSpace'] as string | undefined);
-  if (!wt.ok) { res.status(400).json({ error: wt.error }); return; }
-  const targetSpace = wt.target;
-
-  const result = await bulkWrite(targetSpace, (req.body ?? {}) as BulkInput);
-  if (bulkWriteTotal(result) > 0) {
-    // Bulk suppresses per-item webhooks; emit ONE summary a workflow can inspect.
-    emitWebhookEvent({ event: 'bulk.write', spaceId: targetSpace, entry: { inserted: result.inserted, updated: result.updated, errorCount: result.errors.length }, ...webhookToken(req) });
-  }
-  res.status(207).json(result);
-});
